python/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pyzmq>=22.0.0
python-telegram-bot>=13.0.0
numpy>=1.21.0
Pillow>=8.0.0
ultralytics>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 